Create submit button with two wallets component
